+++ require.php
@@ -2,6 +2,7 @@
 
 require_once 'utils.php';
 require_once 'paths.php';
+require_once 'app.php';
 require_once path_join(LIBS_DIR, 'debug.php');
 require_once path_join(CONFIG_DIR, 'constants.php');
 require_once path_join(FRONTEND_DIR, 'library.php');